requirements.txt
psycopg2-binary==2.9.1
Django==3.2.*
djangorestframework==3.13.*
